Use sync calls for unit test verification.
